% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{add_tags}
\alias{add_tags}
\title{Adds the specified tags to the specified load balancer}
\usage{
add_tags(LoadBalancerNames, Tags)
}
\arguments{
\item{LoadBalancerNames}{The name of the load balancer. You can specify one load balancer only.}

\item{Tags}{The tags.}
}
\description{
Adds the specified tags to the specified load balancer. Each load balancer can have a maximum of 10 tags.
}
\details{
Each tag consists of a key and an optional value. If a tag with the same key is already associated with the load balancer, \code{AddTags} updates its value.

For more information, see \href{http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/add-remove-tags.html}{Tag Your Classic Load Balancer} in the \emph{Classic Load Balancers Guide}.
}
\examples{
# This example adds two tags to the specified load balancer.
add_tags(
  LoadBalancerNames = list(
    "my-load-balancer"
  ),
  Tags = list(
    list(
      Key = "project",
      Value = "lima"
    ),
    list(
      Key = "department",
      Value = "digital-media"
    )
  )
)

}
