% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{delete_load_balancer_policy}
\alias{delete_load_balancer_policy}
\title{Deletes the specified policy from the specified load balancer}
\usage{
delete_load_balancer_policy(LoadBalancerName, PolicyName)
}
\arguments{
\item{LoadBalancerName}{The name of the load balancer.}

\item{PolicyName}{The name of the policy.}
}
\description{
Deletes the specified policy from the specified load balancer. This policy must not be enabled for any listeners.
}
\examples{
# This example deletes the specified policy from the specified load
# balancer. The policy must not be enabled on any listener.
delete_load_balancer_policy(
  LoadBalancerName = "my-load-balancer",
  PolicyName = "my-duration-cookie-policy"
)

}
