% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{describe_account_limits}
\alias{describe_account_limits}
\title{Describes the current Elastic Load Balancing resource limits for your AWS account}
\usage{
describe_account_limits(
  Marker = "string",
  PageSize = 123
)
}
\arguments{
\item{Marker}{The marker for the next set of results. (You received this marker from a previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the current Elastic Load Balancing resource limits for your AWS account.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/elb-limits.html}{Limits for Your Classic Load Balancer} in the \emph{Classic Load Balancers Guide}.
}
\examples{

}
