% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{describe_load_balancer_policy_types}
\alias{describe_load_balancer_policy_types}
\title{Describes the specified load balancer policy types or all load balancer policy types}
\usage{
describe_load_balancer_policy_types(
  PolicyTypeNames = list(
    "string"
  )
)
}
\arguments{
\item{PolicyTypeNames}{The names of the policy types. If no names are specified, describes all policy types defined by Elastic Load Balancing.}
}
\description{
Describes the specified load balancer policy types or all load balancer policy types.
}
\details{
The description of each type indicates how it can be used. For example, some policies can be used only with layer 7 listeners, some policies can be used only with layer 4 listeners, and some policies can be used only with your EC2 instances.

You can use CreateLoadBalancerPolicy to create a policy configuration for any of these policy types. Then, depending on the policy type, use either SetLoadBalancerPoliciesOfListener or SetLoadBalancerPoliciesForBackendServer to set the policy.
}
\examples{
# This example describes the specified load balancer policy type.
describe_load_balancer_policy_types(
  PolicyTypeNames = list(
    "ProxyProtocolPolicyType"
  )
)

}
