% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{remove_tags}
\alias{remove_tags}
\title{Removes one or more tags from the specified load balancer}
\usage{
remove_tags(
  LoadBalancerNames = list(
    "string"
  ),
  Tags = list(
    list(
      Key = "string"
    )
  )
)
}
\arguments{
\item{LoadBalancerNames}{[required] The name of the load balancer. You can specify a maximum of one load balancer name.}

\item{Tags}{[required] The list of tag keys to remove.}
}
\description{
Removes one or more tags from the specified load balancer.
}
\examples{
# This example removes the specified tag from the specified load balancer.
remove_tags(
  LoadBalancerNames = list(
    "my-load-balancer"
  ),
  Tags = list(
    list(
      Key = "project"
    )
  )
)

}
