% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{set_load_balancer_policies_of_listener}
\alias{set_load_balancer_policies_of_listener}
\title{Replaces the current set of policies for the specified load balancer port with the specified set of policies}
\usage{
set_load_balancer_policies_of_listener(LoadBalancerName, LoadBalancerPort,
  PolicyNames)
}
\arguments{
\item{LoadBalancerName}{The name of the load balancer.}

\item{LoadBalancerPort}{The external port of the load balancer.}

\item{PolicyNames}{The names of the policies. This list must include all policies to be enabled. If you omit a policy that is currently enabled, it is disabled. If the list is empty, all current policies are disabled.}
}
\description{
Replaces the current set of policies for the specified load balancer port with the specified set of policies.
}
\details{
To enable back-end server authentication, use SetLoadBalancerPoliciesForBackendServer.

For more information about setting policies, see \href{http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/ssl-config-update.html}{Update the SSL Negotiation Configuration}, \href{http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/elb-sticky-sessions.html#enable-sticky-sessions-duration}{Duration-Based Session Stickiness}, and \href{http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/elb-sticky-sessions.html#enable-sticky-sessions-application}{Application-Controlled Session Stickiness} in the \emph{Classic Load Balancers Guide}.
}
\examples{
# This example replaces the policies that are currently associated with
# the specified listener.
set_load_balancer_policies_of_listener(
  LoadBalancerName = "my-load-balancer",
  LoadBalancerPort = 80L,
  PolicyNames = list(
    "my-SSLNegotiation-policy"
  )
)

}
