% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{delete_load_balancer_listeners}
\alias{delete_load_balancer_listeners}
\title{Deletes the specified listeners from the specified load balancer}
\usage{
delete_load_balancer_listeners(LoadBalancerName, LoadBalancerPorts)
}
\arguments{
\item{LoadBalancerName}{The name of the load balancer.}

\item{LoadBalancerPorts}{The client port numbers of the listeners.}
}
\description{
Deletes the specified listeners from the specified load balancer.
}
\examples{
# This example deletes the listener for the specified port from the
# specified load balancer.
delete_load_balancer_listeners(
  LoadBalancerName = "my-load-balancer",
  LoadBalancerPorts = list(
    80L
  )
)

}
