% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elb_operations.R
\name{describe_load_balancer_attributes}
\alias{describe_load_balancer_attributes}
\title{Describes the attributes for the specified load balancer}
\usage{
describe_load_balancer_attributes(LoadBalancerName)
}
\arguments{
\item{LoadBalancerName}{[required] The name of the load balancer.}
}
\description{
Describes the attributes for the specified load balancer.
}
\section{Accepted Parameters}{
\preformatted{describe_load_balancer_attributes(
  LoadBalancerName = "string"
)
}
}

\examples{
# This example describes the attributes of the specified load balancer.
\donttest{describe_load_balancer_attributes(
  LoadBalancerName = "my-load-balancer"
)}

}
